feat: support forks in `e cherry-pick`
